Update bitcoin.html
https://www.tokens24.com

Tokens24 is an educational and informational website dedicated to helping everyone and anyone learn about cryptocurrency and blockchain technology all in one place. Creating Tokens24 is our way of encouraging greater adoption of this promising area. For newcomers and veterans, Tokens24 is a one-stop-shop for anyone wishing to further their knowledge about cryptocurrencies and blockchain technology.

Tokens24 provides a comprehensive explanation of different cryptocurrencies, as well as an encyclopedia page, that contains extensive definitions on terminology used in the space. Tokens24 is also committed to providing commentary that will allow readers to thoroughly understand cryptocurrency and blockchain technology related concepts that are integral to the space and news from thought leaders in the sector.